Upddate release 1.2.1
Krita 5.2.0 use PyQt 5.15 and Python 3.10, and there's some incompatibilities
to fix.
Bug fix #4

+++ newspaper/newspaper/newspaper.py
@@ -1876,7 +1876,7 @@ class Newspaper(Extension):
 
         dlgMain.hsldSize.valueChanged.connect(lambda v: onHsldSizeChanged(v))
         dlgMain.dspbxSize.valueChanged[float].connect(lambda v: onDspbxSizeChanged(v))
-        dlgMain.hsldSize.setValue(self.__outputOptions['outputSize'] * 100)
+        dlgMain.hsldSize.setValue(int(self.__outputOptions['outputSize'] * 100))
 
         dlgMain.hsldAdjustment.setValue(self.__outputOptions['outputAdjustment'])
         dlgMain.hsldAdjustment.valueChanged.connect(hsldAdjustment_Changed)
